move find_parent_shell to utils
